tslibs.conversions tests
